pyproject.toml
[project]
name = "youtube-to-article-back"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dj-database-url>=2.3.0",
    "django==4.2.9",
    "django-widget-tweaks>=1.5.0",
    "moviepy>=2.1.2",
    "openai>=1.65.4",
    "psycopg2-binary>=2.9.10",
    "pyotp>=2.9.0",
    "python-decouple>=3.8",
    "pytube>=15.0.0",
    "youtube-transcript-api>=0.6.3",
]
